Split up in modules for public repository deprecation of DiscordHook Update config.yml Fixing oversight in Gui registration
