+++ passwords/src/main/java/info/cho/passwords/PasswordsBootstrap.java
@@ -1,0 +1,16 @@
+package info.cho.passwords;
+
+import io.papermc.paper.plugin.bootstrap.BootstrapContext;
+import io.papermc.paper.plugin.bootstrap.PluginBootstrap;
+import io.papermc.paper.plugin.lifecycle.event.types.LifecycleEvents;
+
+public class PasswordsBootstrap implements PluginBootstrap {
+    @Override
+    public void bootstrap(BootstrapContext context) {
+        context.getLifecycleManager().registerEventHandler(LifecycleEvents.COMMANDS, commands -> {
+            // Commands here
+
+
+        });
+    }
+}
