+++ passwords/src/main/java/info/cho/passwords/player/PasswordPlayerMode.java
